refactor(cli): split workers
