+++ cli/ops/worker_host.rs
@@ -5,9 +5,9 @@ use crate::ops::io::get_stdio;
 use crate::permissions::Permissions;
 use crate::program_state::ProgramState;
 use crate::tokio_util::create_basic_runtime;
-use crate::worker::WebWorker;
-use crate::worker::WebWorkerHandle;
-use crate::worker::WorkerEvent;
+use crate::web_worker::WebWorker;
+use crate::web_worker::WebWorkerHandle;
+use crate::web_worker::WorkerEvent;
 use deno_core::error::generic_error;
 use deno_core::error::AnyError;
 use deno_core::error::JsError;
