+++ cli/tools/repl.rs
@@ -4,7 +4,6 @@ use crate::colors;
 use crate::inspector::InspectorSession;
 use crate::program_state::ProgramState;
 use crate::worker::MainWorker;
-use crate::worker::Worker;
 use deno_core::error::AnyError;
 use deno_core::serde_json::json;
 use deno_core::serde_json::Value;
@@ -280,7 +279,7 @@ impl Highlighter for LineHighlighter {
 }
 
 async fn post_message_and_poll(
-  worker: &mut Worker,
+  worker: &mut MainWorker,
   session: &mut InspectorSession,
   method: &str,
   params: Option<Value>,
@@ -305,7 +304,7 @@ async fn post_message_and_poll(
 }
 
 async fn read_line_and_poll(
-  worker: &mut Worker,
+  worker: &mut MainWorker,
   session: &mut InspectorSession,
   message_rx: &Receiver<(String, Option<Value>)>,
   response_tx: &Sender<Result<Value, AnyError>>,
@@ -425,7 +424,7 @@ pub async fn run(
 
   let history_file = program_state.dir.root.join("deno_history.txt");
 
-  post_message_and_poll(&mut *worker, &mut session, "Runtime.enable", None)
+  post_message_and_poll(&mut worker, &mut session, "Runtime.enable", None)
     .await?;
 
   // Enabling the runtime domain will always send trigger one executionContextCreated for each
@@ -474,7 +473,7 @@ pub async fn run(
 
   while !is_closing(&mut worker, &mut session, context_id).await? {
     let line = read_line_and_poll(
-      &mut *worker,
+      &mut worker,
       &mut session,
       &message_rx,
       &response_tx,
@@ -495,7 +494,7 @@ pub async fn run(
         };
 
         let evaluate_response = post_message_and_poll(
-          &mut *worker,
+          &mut worker,
           &mut session,
           "Runtime.evaluate",
           Some(json!({
@@ -513,7 +512,7 @@ pub async fn run(
             && wrapped_line != line
           {
             post_message_and_poll(
-              &mut *worker,
+              &mut worker,
               &mut session,
               "Runtime.evaluate",
               Some(json!({
@@ -533,7 +532,7 @@ pub async fn run(
 
         if evaluate_exception_details.is_some() {
           post_message_and_poll(
-                    &mut *worker,
+                    &mut worker,
                     &mut session,
                     "Runtime.callFunctionOn",
                     Some(json!({
@@ -546,7 +545,7 @@ pub async fn run(
                   ).await?;
         } else {
           post_message_and_poll(
-                    &mut *worker,
+                    &mut worker,
                     &mut session,
                     "Runtime.callFunctionOn",
                     Some(json!({
@@ -564,7 +563,7 @@ pub async fn run(
         // Deno.inspectArgs.
         let inspect_response =
           post_message_and_poll(
-            &mut *worker,
+            &mut worker,
             &mut session,
             "Runtime.callFunctionOn",
             Some(json!({
